% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henkiloverotus.R
\encoding{UTF-8}
\name{henkiloverotus_2025}
\alias{henkiloverotus_2025}
\title{Henkilöverotus vuodelle 2025}
\usage{
henkiloverotus_2025(
  aineisto,
  ika,
  lapsi_lkm_yksin,
  lapsi_lkm_yhteis,
  kunnan_veropros,
  kirkko_veropros,
  elaketulo,
  palkkatulo,
  etuustulo,
  pyoristys = T
)
}
\arguments{
\item{aineisto}{data-frame}

\item{ika}{henkilön ikä vuoden lopussa}

\item{lapsi_lkm_yksin}{yksin huollettavien lasten lukumäärä}

\item{lapsi_lkm_yhteis}{yhteishuollettavien lasten lukumäärä}

\item{kunnan_veropros}{kuntaveroprosentti (esim. 0.05)}

\item{kirkko_veropros}{kirkollisveroprosentti (esim. 0.01)}

\item{elaketulo}{elaketulo vuoden aikana}

\item{palkkatulo}{palkkatulo vuoden aikana}

\item{etuustulo}{etuustulo vuoden aikana}

\item{pyoristys}{Kuvien piirtämistä varten lukujen pyöristyksen voi ohittaa}
}
\value{
Verotuksen tulos
}
\description{
Henkilön palkka-, etuus- ja eläketulon verotus vuonna 2025
}
\examples{
# 100 tapausta yksi per data.frame-rivi
n <- 100

# generoidaan satunnaista testidataa
aineisto<-data.frame(
ika = sample(seq(16,70),n,replace=TRUE),
lapsi_lkm_yksin = sample(c(0,1),n,replace=TRUE),
lapsi_lkm_yhteis = sample(c(0,1),n,replace=TRUE),
kunnan_veropros = round(runif(n,0.15,0.25),3),
kirkko_veropros = round(runif(n,0.00,0.01),3),
elaketulo = round(runif(n,0,10000),0),
palkkatulo = round(runif(n,0,50000),0),
etuustulo = round(runif(n,0,10000),0)
)

# aineiston voi valitaa parametrina
tmp2 <- henkiloverotus_2025(aineisto,ika,lapsi_lkm_yksin,lapsi_lkm_yhteis,
  kunnan_veropros,kirkko_veropros,elaketulo ,palkkatulo,etuustulo)

# tai vaihtoehtoisesti toimii myos pipe-operaattori
tmp1 <- aineisto |> henkiloverotus_2025(ika,lapsi_lkm_yksin,lapsi_lkm_yhteis,
  kunnan_veropros,kirkko_veropros,elaketulo,palkkatulo,etuustulo)

}
